# lib/openapi/spec.json
{"openapi":"3.1.0","info":{"title":"ПАПА API","version":"1.0.0","description":"Программа автоматизации производственной аналитики"},"servers":[{"url":"/","description":"Base URL"}],"paths":{"/api/health":{"get":{"summary":"Health check","description":"ALB health check, no auth","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"ok":{"type":"boolean"}},"required":["ok"]}}}}}}},"/api/audit/events":{"get":{"summary":"List audit events","security":[{"session":[]}],"parameters":[{"name":"action","in":"query","schema":{"type":"string"}},{"name":"actor","in":"query","schema":{"type":"string"}},{"name":"from","in":"query","schema":{"type":"string","format":"date"}},{"name":"to","in":"query","schema":{"type":"string","format":"date"}},{"name":"limit","in":"query","schema":{"type":"integer","default":50}}],"responses":{"200":{"description":"Events list"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"summary":"Log audit event","security":[{"session":[]}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"action":{"type":"string"},"target":{"type":"string","nullable":true},"metadata":{"type":"object"}},"required":["action"]}}}},"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"}}}},"/api/admin/users":{"get":{"summary":"List users","security":[{"session":[]}],"responses":{"200":{"description":"User list"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/compliance/inbox":{"get":{"summary":"Compliance inbox","security":[{"session":[]}],"responses":{"200":{"description":"Inbox items"},"401":{"description":"Unauthorized"}}}},"/api/compliance/monitor":{"get":{"summary":"Compliance monitor","security":[{"session":[]}],"responses":{"200":{"description":"Monitor data"},"401":{"description":"Unauthorized"}}}},"/api/inspection/cards":{"get":{"summary":"Inspection cards","security":[{"session":[]}],"responses":{"200":{"description":"Cards list"},"401":{"description":"Unauthorized"}}}},"/api/mail/inbox":{"get":{"summary":"Mail inbox","security":[{"session":[]}],"responses":{"200":{"description":"Mail items"},"401":{"description":"Unauthorized"}}}},"/api/anchoring/health":{"get":{"summary":"Anchoring health","description":"Pre-login dashboard, no auth required","responses":{"200":{"description":"Health status","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["OK","DELAYED","FAILED","UNAVAILABLE"]},"network":{"type":"string"},"chainId":{"type":"integer"}}}}}}}},"/api/compliance/break-glass":{"get":{"summary":"Break-glass status","security":[{"session":[]}],"responses":{"200":{"description":"Status"},"403":{"description":"Forbidden"}}},"post":{"summary":"Activate/deactivate break-glass","security":[{"session":[]}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"action":{"type":"string","enum":["activate","deactivate"]},"reason":{"type":"string"}}}}}}},"responses":{"200":{"description":"OK"},"403":{"description":"Forbidden"}}}},"/api/compliance/policies":{"get":{"summary":"List policies","security":[{"session":[]}],"parameters":[{"name":"key_class","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Policies"},"403":{"description":"Forbidden"}}},"post":{"summary":"Create/update policy","security":[{"session":[]}],"responses":{"200":{"description":"OK"},"403":{"description":"Forbidden"}}}},"/api/ledger/verify":{"get":{"summary":"Verify ledger chain","security":[{"session":[]}],"responses":{"200":{"description":"OK"},"403":{"description":"Forbidden"},"409":{"description":"Chain integrity failed"}}}},"/api/proof/artifact":{"get":{"summary":"Events by artifact SHA-256","security":[{"session":[]}],"parameters":[{"name":"sha256","in":"query","required":true,"schema":{"type":"string","pattern":"^[a-f0-9]{64}$"}}],"responses":{"200":{"description":"Events"},"400":{"description":"Invalid sha256"},"403":{"description":"Forbidden"}}}}},"components":{"securitySchemes":{"session":{"type":"apiKey","in":"cookie","name":"next-auth.session-token"}}}}